+++ src/main/java/admin/adminsiteserver/common/response/PageInformation.java
@@ -1,4 +1,4 @@
-package admin.adminsiteserver.common.dto;
+package admin.adminsiteserver.common.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -6,15 +6,15 @@ import org.springframework.data.domain.Page;
 
 @Getter
 @AllArgsConstructor
-public class PageInfo {
+public class PageInformation {
     private int currentPage;
     private int totalPages;
     private int numberOfElements;
     private int pageSize;
     private int totalElements;
 
-    public static <T> PageInfo from(Page<T> page) {
-        return new PageInfo(
+    public static <T> PageInformation from(Page<T> page) {
+        return new PageInformation(
                 page.getNumber(),
                 page.getTotalPages(),
                 page.getNumberOfElements(),
