+++ src/main/java/admin/adminsiteserver/qna/ui/QnaController.java
@@ -1,6 +1,6 @@
 package admin.adminsiteserver.qna.ui;
 
-import admin.adminsiteserver.common.dto.CommonResponse;
+import admin.adminsiteserver.common.response.CommonResponse;
 import admin.adminsiteserver.authentication.ui.AuthenticationPrincipal;
 import admin.adminsiteserver.authentication.ui.LoginUserInfo;
 import admin.adminsiteserver.qna.application.QnaService;
